move HotkeyAction to api.features and merge it with HotkeyActions rename HotkeyActionsImpl to HotkeyActions
