+++ src/main/java/appeng/hotkeys/InventoryHotkeyAction.java
@@ -6,7 +6,7 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 
-import appeng.api.hotkeys.HotkeyAction;
+import appeng.api.features.HotkeyAction;
 
 public class InventoryHotkeyAction implements HotkeyAction {
